Extract yield and next types from generator return types early
Summary:
Currently, the next and yield types for generators are open tvars that accumulate bounds during typechecking of generator function bodies.

To begin making this work compatible with local type inference, this diff extracts out the yield and next types from the overall function return type ahead of time. To do so, we need to flow a `Generator` containing tvars into the return type (which we expect, if annotated, to have a form like `Generator<Yield, Return, Next>` or `Iterator<Yield>`) to receive lower bounds on next and upper bounds on yield, and then flow the return into `$iterable` to obtain a lower bound on the yield (because the variance on generators is `<+Yield, +Return, -Next>`. We can then populate the environment with these types.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D36534490

fbshipit-source-id: ab9516871815db522cff0bae568325c54f1d8bdc