Treat generator next as a parameter-like entry in the env builder
Summary:
The `next` of a generator operates essentially like a function parameter. It's the type that is passed into the generator when `.next` is called on it, and internally the expression `yield` (with or without argument) has the `next` type.

We can model this nicely in the env builder by treating all `yield` expressions as reads of a special `yield` variable, whose write location is the return annotation (since that's where the `next` is annotated) of the enclosing generator function.

We can then populate this entry in the new env directly when typechecking function bodies, at the point that we've extracted the next type from the return annotation.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D36685423

fbshipit-source-id: 9e68ac5a87b4ea623617ea1ba06dadd383efcc15