Create t_out placeholders in yield and return
Summary:
This diff creates a `'T` type placeholder in the AST nodes for `yield` expressions and `return` statements. This type represents the actual value type that is yielded or returned (which can be different from the type of the argument to yield or return: a `yield*` expression's yield type is the yield type of the delegated-to expression, and the return type of a predicate function is an OpenPredT.

This "slot", populated in statement.ml, will be used in subsequent diffs.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D36686465

fbshipit-source-id: 0322d68edc36a9c0c7f5b9b40d1bb205f9f6b850