[new-env] Define instance this and static this at class level
Summary:
Changelog: [internal]

With all the preparations and refactors in previous diffs, we are finally ready to make a lot of `this` definitions per class. This diff follows similar setup for per-class super in D36682132, and it's mostly mechanical thanks to the work already done in D36682132.

There are only a few interesting changes:
1. We no longer need to introduce `this` bindings to class properties and class private fields.
2. When in a class method, we will only introduce additional this binding for `this` annotation.

Reviewed By: mvitousek

Differential Revision: D36709081

fbshipit-source-id: aa0f1b018ef855cf882087fbbec7bbd28b35ddc1