Replace `this_recipe` with `default_this`
Summary:
Changelog: [internal]

After a series of refactors in previous diffs, it's becoming clear that a non-null `this_recipe` only appears in two places: `class_sig.ml` and `mk_function` in `statement.ml`, and they both perform the following computation:

```
let t = this_param fparams |> annotated_or_inferred_of_option ~default in
(type_t_of_annotated_or_inferred t, t)
```

The only difference is on how to compute the default. Therefore, it makes sense to only pass in the default value, and left the rest to be handled by `func_sig`.

While we can make the default value of type `Type.t option`, we can actually do better. I created a new type with 3 variants:

```
type default_this_type =
    | ParentScopeThis
    | ClassThis of Type.t
    | FunctionThis of Type.t
```

This makes the code self-documenting and easier to reason about why we store or not store `this` into env. This will also make it easier to work with the new-env, since the extra tag information that distinguishes between class this and function this allows the new-env to decide whether to bind a new `this`.

Reviewed By: mvitousek

Differential Revision: D36685092

fbshipit-source-id: e3c951ecc8af9944c75f07ac76300544c460d460