Make `this_recipe` itself optional instead of returning optional types from `this_recipe`
Summary:
Changelog: [internal]

Right now, `this_recipe` is created in 3 places:
1. `class_sig.ml` when type checking each methods
2. `mk_function` when type checking a function created with `function` syntax
3. `make_arrow` when type checking an arrow function

In 1 and 2, `this_recipe` always return `(this_t, this)`, where `this` is stored in the env, and `this_t` is passed back from `Func_sig.toplevels` to create a function type.

`make_arrow` is the special case here. Arrow functions do not create their own `this` types, so it returns `None` for `this` to indicate that, and that's the only reason we are returning optional types from `this_recipe`.

This diff refactors `this_recipe`, so that it will never return optional types but instead making the function itself optional. Most of the changes in this diff is caused by changing where we construct options.

Reviewed By: mvitousek

Differential Revision: D36684077

fbshipit-source-id: 5bbd59370c92a109c586e6a3c7f533f15d4844d0