[resolved-env] Do not consider contextual binding to be resolved when contextual typing flag is not on
Summary:
Right now we consider the contextually typed unannotated parameters to be resolved without contextual typing, and assign it an open tvar. Since it's resolved, no bounds can be added to it, so the tvar will act as `any`.

This diff prevents the locations of unannotated parameters to be added to the resolved set, so that we can have resolved-env roll out first.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D36781419

fbshipit-source-id: 3c52a72fd02a38c226d01b981636355641b1d25c