Check function return and yield after checking function body
Summary:
This diff removes the dependency on the old env for `yield` and partially removes it for `return`. Rather than these types existing in the environment, we now only populate the `out` slots in the typed ast (defined in the previous diff) during typechecking of yield and return statements. Then, once we've finished checking the body, we do an additional pass over the body in the typed AST--this pass extends the existing pass used to determine exhaustiveness. We can then see the returned/yielded type from each node, and flow them into the overall return/yield type in whatever the appropriate way is (as a promise for async functions, a generator for generator functions, etc).

The one way we still rely on the old env is that we still need to get a hint when typechecking the return expression. This is handled separately in the next diff.

Reviewed By: SamChou19815

Differential Revision: D36686466

fbshipit-source-id: 47c1a73020023da4908b73fa832d35430ba0bad2