[contextual] Make RHS have hint of decomposition of object prop in `obj.prop = RHS`
Summary:
Changelog: [internal]

This diff passes a proper hint to the RHS in member assignments. Note that we can't use the type of LHS directly, since LHS's type is decided by the RHS. Instead, we need to go with the indirection of `Decomp_ObjProp prop` with type of `obj` in `obj.prop` as the base.

Reviewed By: mvitousek

Differential Revision: D36788357

fbshipit-source-id: 6001dd73034831e0c6891cdd03ad2d3e3e61f47c