Include return hints in loc env
Summary:
To finally break the dependence on the old env, we need to provide a return type hint to `return` statements. The simplest way to do so is as part of the context (via, in this case, the loc env). We simply populate a slot in the loc_env with the return hint before entering function scopes and pop it off at the end.

I believe that this is compatible with the fully resolved environment as currently designed, because the FRE will not typecheck return statements out of order. (One possible exception is a return statement like `return (x = { foo: 42 })`, where we could potentially check the assignment expression out of order and lose a hint applied to it, but this feels like a pretty remote possibility.)

There is a slight difference in the behavior of annotation requirements. Previously we were always extracting hints from return types, but those hints could be wrong in the case of async or generator functions, where the argument to `return` shouldn't actually correspond with the type of the annotation. We could use type decomposition operators on the annotation (if powerful enough ones are available) to figure out the hint based on the annotation, but in this case we just leave async and generator returns without hints.

Reviewed By: SamChou19815

Differential Revision: D36710060

fbshipit-source-id: d9bf6f81b73b7d7cf45095fe2cad4193ec564aa6