Pass scope kinds through to resolved environment
Summary:
In a few places (delegate yields, function returns, toplevel type annotations, constructors) we need to know what kind of scope we're in to decide how expressions or statements should be typechecked. This diff implements that in the fully resolved environment, by adding scope kind tracking to the name_def analysis and including it as part of all name defs. When we resolve a name def, we first set the scope kind appropriately.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D36750229

fbshipit-source-id: a49ba2e0f99918e1017cda8509725a5f73c3d6e7