Restrict unannotated generator next
Summary: Currently, the `next` field of a generator, if the generator is unannotated, is an open tvar, which is populated by calling `.next(x)` on the generator object. This is trivially incompatible with local type inference, of course. If a generator is annotated then there's no problem, but we need to choose a behavior for unannotated ones. This diff makes unannotated generator `next`s be `void`, which matches the behavior of `Iterator`s.

Reviewed By: SamChou19815

Differential Revision: D36727896

fbshipit-source-id: 73c493bdb73ca44b8ffa4c438672e0d42b724aa0