Loosen toplevel/global env checking to allow for lexical scopes
Summary:
In coming diffs we're going to use the new env to track scope kinds rather than depending on the old env for this. Currently, the `is_toplevel_scope` and `is_global_scope` predicates are the only place that typechecking contextually cares about whether we're currently in a lexical scope or not: they return `false` if we're in a lexical scope directly over the toplevel/global scope. All other scope queries only care what the closest var scope is.

Having to worry about lexical scopes here greatly complicates the work of figuring out the scope context in the new env, and it turns out to be safe to loosen these predicates to return true if they're in lexical scopes: the toplevel check is used to determine if we should use source maps or generated maps for type annotation IDs, the only difference being that source maps are slightly slower to compare but are shared with the merge phase, and `is_global_scope` is only relevant in libdefs which shouldnt have lexical scopes anyways.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D36747880

fbshipit-source-id: 16c14c1831e52590e97428b77b0273b9b667eb03