Populate loc_env with class bindings and use for private member lookup
Summary:
Currently the env is used to power private property lookup: whenever we see a private access, we grab the current stack and find whatever classes we're currently within the body of. We can then look through these classes until we find one whose id matches the id of the receiver of the private lookup, and look it up on its private methods.

This diff implements that for the new env, breaking another dependence on the old env. We do so by maintaining a map of class_bindings in the loc env, keyed by the location of the class, which we update as we generate class sigs. We also maintain a stack of class locations to indicate which classes we're currently inside the scope of. When we see a private lookup, we just use this contextual stack to generate the list of class bindings, and proceed as usual.

Why do we need both a location-keyed map, and a contextual stack? Why don't we just have a stack of class_bindings? This is to support the fully resolved environment: the location stack can be computed before typechecking in the name_def pass, and will need to be provided as part of resolution for each name to make sure that we populate the current stack correctly before resolving a name. This will be implemented subsequently, but the design in this diff is intended to support that.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D36779445

fbshipit-source-id: 3db130c6c946e78622498cbeb36ad0fd038ef9d1