Track scope kind in loc env
Summary:
In order to fully break the dependence on the old env, the new env needs to know what kind of scope it is in at any point. The natural place to track this is within the loc env.

For the purposes of the "regular" new env--not the fully resolved environment--we can track this by using the existing `push/pop_var_scope` functions to update the stored var scope in the new env. To avoid having to build a stack within the loc env (which wouldnt make much sense when we extend this to the FRE) we have the callers of push/pop_var_scope handle reinstating the old scope kind afterwards.

We can then directly answer predicates like `in_predicate_scope` from the new env rather than deferring to the old env.

The FRE is handled separately in a later diff.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D36724562

fbshipit-source-id: 382944585663f1a28da5e6b4edc812e4e1a81b24