[parser] check for duplicate exports in a post-parsing phase
Summary:
duplicate exports are parse errors, so we check for them in the parser. it is entangled in individual parsing functions and mutates the env in place. it's hard to be sure we covered all the cases, especially in error cases. there are some `failwith`s that keep getting hit, causing runtime crashes.

instead, this should be a post-parsing analysis. I kept it in the parser (for now) but moved it into a single `check_for_duplicate_exports` function that we call from `program` (for now).

this fixes a crash when doing `export async ...` where `...` is not `function`, but causes us to incorrectly parse that without errors. the next diff will fix that.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D36271857

fbshipit-source-id: 385b4c71a6d80d844b48e554850ef118cf63ce3f