improve error recovery for invalid export declarations
Summary:
the catch-all case for parsing export declarations expects `export {`. error recovery for invalid stuff is very poor, because it uses `Expect.token` and keeps consuming error tokens.

this diff changes to erroring on the unexpected token and dropping the `export` from the ast. the `export` is like a modifier, so I think it's more likely that someone added an `export` in the wrong spot and what follows is valid-ish, than that they meant a `{`.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D36449167

fbshipit-source-id: 6f15988a338e5adfb2faff120154929597300b51