Scopes for private members in fully resolved env
Summary:
This diff adds support for private member lookup in the FRE. It does so by building the class stack during the name_def traversal pass, and associating the current class stack with all name defs. When we typecheck a name def, we simply need to install the appropriate class stack in the loc_env, and then any private properties that are looked up within the def will be available.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D36790490

fbshipit-source-id: d3ede92e9ac2c6c3e96023178294f5caf0dda91c