fix Type_sig exception on invalid qualified types
Summary:
if a qualified type annot is reachable from an export, which references a generic, then `Type_sig_parse` currently hits an assertion and raises.

this breaks IDE services as long as that pattern is in the file.

the smallest repro I found is:
```
export type T<U> = U.V
```

but a repro I found in the wild was more like:
```
type Props<T> = { value: T.U }
export function Foo(props: Props<T>): ... {}
```

this is a type error -- "Cannot resolve name `T`. [cannot-resolve-name]" and "Cannot use `U` as a type because it is an `any`-typed value." (??) -- so I used `CheckError` to make it `any` and not raise.

Changelog: [fix] Fix broken IDE services when the file contains a certain invalid type annotations

Reviewed By: SamChou19815

Differential Revision: D36928654

fbshipit-source-id: fe72d164d31611b8f7e241a631308f3b7e2fdbbc