[contextual] Add two interesting tests on decomposting a generic function hint
Summary:
Changelog: [internal]

In Pierce's 2000 LTI paper, we must have all arguments fully annotated when the callee is polymorphic. This is completely unacceptable to us, since it would a lot more annotations in places people don't expect to add. e.g. on lambdas passed to map functions:

```
declare function map<R>(arr: Array<string>, f: (string) => R): Array<R>
map(stringArr, (s) => s);
```

Therefore, we must relax this rule. We can use a two pass approach: in the first pass, we use some arguments that can be type checked independently as part of the constrains to solve type arguments as much as we can. Ideally and I believe practically, this should leave us with a monomorphic function type at the end, that we can use to contextually type the lambda terms. However, this is not a guarantee. A simple example would be

```
declare function f<T>(T => void) => void;
f((a) => {});
```

In this example, there are no constrains we can add initially to solve for type arguments, so we are forced to decompose a generic function type hint. Therefore, it's interesting to see what will happen in this case.

In this diff, I added two test cases:
- The first case represents the case when we try to decompose something that's not a function, and it fails as expected.
- The second case represents the case that we are still decomposing a function, but the function is generic. In this case, we produced the `mixed` type for unsolved type arguments, which I think is acceptable.

Reviewed By: panagosg7

Differential Revision: D36898536

fbshipit-source-id: 4650ee586c2c29adf3688e76643796f2158929a7