more descriptive sharedmem exception
Summary:
we're seeing rare exceptions where we try to read a header from sharedmem and it's already been freed.

this diff adds a specific exception and reraises it in the primary place we've seen this issue, so we can see which file is corrupt. hopefully then we can correlate it with what else happened -- e.g. was it a deleted file?

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D37014516

fbshipit-source-id: c2a0b7d114269bd67b8e84bebd5fcdefd8645c48