[contextual] Use separate `hint_node` type
Summary:
Changelog: [internal]

When I first introduced the integration between contextual typing and fully resolved environment, I decided to reuse the root type as the type of the base hint, since it happens to contain two variants (annotation and value) that I need.

However, this turns out to be very awkward. Resolving the base hint is usually much simpler than resolving a root, since all we need for hint is a type. In addition, when I tried to add another variant `Union` (done in the next diff) that tries to encode a union of types for the purpose of assignments hint using providers, it becomes even worse, since the union root is only designed for hint, but for roots.

Therefore, in this diff, I separate these two concepts and use a separate `hint_node` type specifically for the contextual root.

Reviewed By: panagosg7

Differential Revision: D36831461

fbshipit-source-id: 2db1bb9989e236db89ec2c803a7c0ce76411f97e