Codemod to add annotations with optional properties on prop-missing errors
Summary:
This is mostly based on Daniel's original diff D30847830

I have updated it to work with the changes to the codemod infra since he wrote it. Additionally, I added proper reporting of validation errors, and removed adding annotations for `{}` as an expression (outside of variable declarations) as I haven't found that useful.

Reviewed By: SamChou19815

Differential Revision: D36939876

fbshipit-source-id: 91f29c46d2f55318d9ae8a5468398431b32884fd