fix flatten_spread_args on empty array literals
Summary:
When spreading an array `[...a]`, we compute a flattened version of the "elements"
of `a`  in `flatten_spread_args`. This set of elements is then propagated to the type
of the array element of the result.

The way that this set of elements is computed depends on the kind of array `a`.
If the input is a tuple, then each one of these elements are considered separately.
Array literals types are considered as tuples. The contents at each position
of the tuple are determined by the second field of `ArrayAT`.

This logic relies on the fact that the general type of the array (first field) is equivalent to
(the union of) the set of each one of its elements. This logic, however, breaks down for
the case of the empty array literal, where the tuple list is empty. In that case,
`flatten_spread_args` would return an empty list, and so no information would be
propagated to the result, when instead the general type of the array (to be inferred)
should be considered. The reasoning for this is to account for cases like
```
var a = [];
a.push(1);
var b: Array<string> = [...a];
```
where the general type of the literal is determined later on through a `.push()`.

This diff fixes this by handing specially the case that corresponds to the empty literal
`ArrayAT (_, Some [])`, and returning an element that corresponds to the general
type instead.

Changelog: [fix] fixes unsound spreading of empty array literals (e.g. [try-Flow](https://flow.org/try/#0G4QwTgBCELwQ2gXQNwCgQDoAOBXAzgBYAUAjAJRqiQBGAXBAIJhggCeAPHgC5gCWAdgHMAfLAQYJIFBAD0MiIQD2 (https://github.com/facebook/flow/commit/22839cdfed64a3640dc61de4c8bb2395999c0113)OADYATCAFNmisEA))

Reviewed By: SamChou19815

Differential Revision: D37057875

fbshipit-source-id: 787205c2386de0c8cbddc13dd3da4fd35c29b60e