[resolved-env] Actual support for member assignments in fully resolved environment
Summary:
Currently, the support for member assignments and updates in fully resolved environment are very broken. They are never considered as a write, so all of their write slots are either always populated with an unconstrained tvar or happened to have some write type, because the member expression is refined. This diff adds actual support for member assignments.

For op assignments and updates, it's relatively more straightforward. We only need to add the case when the LHS is a member, and record the LHS node in the def, so that it can later be type checked in `env_resolution`.
For assignments without op, we need to be careful that we don't read the entire member expression, to avoid recording a refined expression or chain expression. Therefore, we use `super#member` so that we will only visit subexpressions of the member expression.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D36784953

fbshipit-source-id: dc789785cbbf416681c6f9f765809877511c9be4