[resolved-env] Pass cond to chain expression (1/n)
Summary:
Changelog: [internal]

The cond parameter for type checking member expressions decides whether we are in normal checking mode or we are in a conditional context. It decides whether we use `GetPropT` or `TestPropT`.

Currently, we have no contextual information in `env_resolution` to decide whether the `cond` parameter should be, so we always pass in `None`. This diff fixes that by passing in the correct `cond` parameter in most cases. This diff is not exhaustive. There are still a lot of unhandled cases like `return` and various refinements in eq_test.

Reviewed By: mvitousek

Differential Revision: D36977513

fbshipit-source-id: 35aef9408b7ca59410061fc339b16b0a69be3a61