[resolved-env] Fix tparams support in resolved env
Summary:
Currently, tparams are modeled incorrectly in a few different ways:

1. Sometimes crashes when tparams are in a SCC. (See explanation at the end)
2. Dependency on tparams are not reported correctly in name_def_ordering
3. tparams are not resolved with an empty tparams_map, causing it to miss some bounds and defaults.

This diff fixes these problems by

1. Do not crash when there are no entry for tparam in resolved-env. Instead, default to any.
2. Add dependencies on tparams in name_def_ordering
3. Attach tparams_map to the tparams def

## Appendix: Explanation on problem 1

Right now, we resolve a type param without any tparams map. This is wrong, since type params might refers to other type params in default and bound. Therefore, similar to annotations, we also need to attach the currently in-scope tparams.

Following the steps described above, we got a lot of new crashes when we couldn't find tparams corresponding to a given loc in `Loc_env.tparams`. So what's happening there? It turns out that there is no guarantee we will resolve tparams in the order declared in the source, and there is even no guarantee that we will resolve the tparams before the function/class that defines it. To see why, let's consider the following example:

```
class Foo<A: Bar, B: Bar> {}
type Bar = Foo<Bar, Bar>
```

The dependency graph will be:

```
-> Bar -> A --
|  |         |
|  |         \/
|  --> B -- Foo
|            |
|            |
--------------
```

We have a strongly connected component, and a one we consider legal (doesn't even matter whether it's legal or not, since we will still visit illegal ones in the same way), because the edges are between annotations. Since it's a SCC, we have no guarantee of the visiting order, and everything is at the mercy of the tie breaking mechanism in topological sort.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D36912180

fbshipit-source-id: 8584dd1b2015824c382318c747966f454823551f