module.exports in new_env
Summary:
This builds support for `module.exports` in the new env, breaking the dependency on the `InternalName "exports"` variable in the old env.

Essentially, this diff removes special-casing for `module.exports` in statement.ml when the new env is enabled. We instead track refinements and updates to it normally using the name resolver, which now is aware of `module.exports` as a pre-existing heap refinement on `module`.

We still need to special-case writes to module.exports slightly. In `module.exports = foo`, type-at-pos on `module.exports` would have the type of `foo`, which is a difference from how typical prop updates would work--normally we would record the type that `x.y` has normally. But in this case, that's `any`, so being consistent there would greatly sabotage coverage without providing anything actionable. So in this case, we maintain the old behavior of the typed ast while removing the special case from typechecking itself.

Reviewed By: SamChou19815

Differential Revision: D36980426

fbshipit-source-id: 98cd41d60eeb98a4bbc9795f98874a51410461fc