[output] Change list_with_newlines to default to no hard newline for new nodes
Summary:
`list_with_newlines` utility is used to add/preserve extra newlines based on original source locations e.g.
```
function A(
  a: string,

  b: number
) {}
```
It does this by comparing the loc line of the previous item in the list to see if they are greater than the current node's loc line by `1`. However for new nodes where the loc is `Loc.none` properties are `0` then this is always true, so the default for new nodes is to force a double newline even though this is rare in actual code. This change identifies this case and defaults to no forced line break to closer match prettier's output.

Changelog: [internal]

Reviewed By: bradzacher

Differential Revision: D37125520

fbshipit-source-id: c02f4188c858be7d6fd6430b85d0323e27919d56