fix race leading to dropped workloads
Summary:
previously, we did this to wait for the next "workload" (LSP/CLI commands):

```
    let%lwt workload =
      Lwt.pick
        [
          ServerMonitorListenerState.wait_and_pop_parallelizable_workload ();
          (let%lwt () = wait_for_cancel in
           raise Lwt.Canceled
          );
        ]
```

`Lwt.pick` chooses the first thread to resolve and cancels the other. but if they both resolve simultaneously, it chooses the one that failed (the cancel thread in our case) and throws out the result of the other.

since `wait_and_pop_parallelizable_workload` pops the workload, throwing out the result means we never respond to that workload!

instead, only wait for a workload. after we're no longer at risk of being cancelled, then we can pop the workload and respond to it.

NOTE: this fix is mostly theoretical based on reported symptoms: the logs would show "Enqueueing parallelizable ..." and "Dequeueing parallelizable ..." for a given request, but then never say "Running a parallel workload" or "Persistent request: ..." or "Persistent response: ...", indicating that `WorkloadStream.pop_parallelizable` was being called but the workload was never making it past this `Lwt.pick`.

Changelog: [fix] Fix a race condition that could cause individual requests from the IDE to hang indefinitely

Reviewed By: jbrown215

Differential Revision: D37184070

fbshipit-source-id: f48d0eed59276b52078c888c6911c809c40dd779