add flag to disable restarting to save time
Summary:
we record a bunch of stats about how long each recheck takes, so that we can estimate how long the next recheck will take. we restart if it's estimated to take longer to recheck than to restart.

this constantly gets in my way when trying to repro bugs. my recheck estimates are different than the user's, so my server restarts when the user's didn't, making it impossible to debug incremental issues.

this diff adds a `--estimate-recheck-time=false` flag to `flow server` and `flow start`, as well as a `FLOW_ESTIMATE_RECHECK_TIME=false` env.

I also added an `estimate_recheck_time=false` flowconfig option for completeness, and so we can maybe even experiment with disabling this and see if it's really helping.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D37248849

fbshipit-source-id: c27567ea84e317742de1616cdf93f2e629eb3731