add --saved-state-skip-version-check-DO_NOT_USE_OR_YOU_WILL_BE_FIRED flag
Summary:
we go to great lengths to make sure that a saved state was generated with the same exact binary that is loading it, because it's just marshal'd ocaml data and really easy for random diffs to change the format. reading it back is completely untyped and leads to segfaults if it's different.

however, sometimes we need to add debug logging or something like that, but still test with saved state. this diff adds a `--saved-state-skip-version-check-DO_NOT_USE_OR_YOU_WILL_BE_FIRED` flag we can use for this VERY LIMITED purpose.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D37252410

fbshipit-source-id: 74f1c7255c8008eddbd987d3b338e1f9dd3e464e