fix watchman restart recovery
Summary:
when attempting to recover from watchman restarts, we tried to query for the mergebase without first calling `watch-project`. since watchman just restarted, it's possible nothing has watched it yet, which makes the mergebase query fail repeatedly, and we fail to recover. further, we're trying to query a dead socket which leads to an EBADF that we have to recover from as well.

instead, we should close the dead socket and re-init the connection, which includes calling `watch-project`.

Changelog: [fix] Improve ability to recover from Watchman restarting

Reviewed By: jbrown215

Differential Revision: D37279786

fbshipit-source-id: 5c503c12ed107503e338e82975730d6a0856b220