[resolved-env] this typeparam support in resolved-env
Summary:
Changelog: [internal]

Based on the initial work done in D37199436, we can finally add support for `this` tparam in fully resolved environment. Unlike other defs in FRE, this param has no syntactic def locations, and there is no environment entry for this param. Instead, what we want to resolve is its tparam entry in `Loc_env`.

The implementation is mostly straightforward. We pass relevant info from name_def, and construct the this tparam in the same way we construct it in `statement.ml` and `class_sig`. What's more interesting is the implications of the change.

Consider any named/nameless class `class { ... }`, its signature is modeled like after the change in D37199436:

```
type self = class_sig <this: self> { ... }
```

The change in this diff implicitly creates a cycle between `this` and `self`, both non-syntactic. This diff added `this` (the implicit type param) as a def, but `self` remains an entry in the env (added in D37199436) that is not a def in resolved-env yet. It shouldn't be too hard to add this in the future, but just like `this` and `super` values, not adding them now doesn't seem to break anything.

Reviewed By: jbrown215

Differential Revision: D37213875

fbshipit-source-id: c0fe592979fa2b3d9a187e6a7e7d22b8848b5db6