[resolved-env] Consider predicate functions and functions with unannotated params as inferred
Summary:
Changelog: [internal]

Currently, the `fully_annotated` field of the `Function` def variant is terribly named. It considers functions with unannotated parameters as "fully annotated" as long as return type is annotated. It also considers predicate functions as fully annotated even though the real return type of predicate functions come from the body. It turns out this classification runs into real problems.

For predicate functions, it's obvious that the return annotation doesn't really dictates its behavior when being called. We need to type check the return in conditional context to figure out the real return type. Therefore, it should be considered to be `inferred` by `env_resolution`.

For functions with unannotated parameters, we have a different problem. Consider the following program:

```
function bar(i): number {
  declare var a: Array<string>;
  (a[i]: number); // missing error
}
bar(0);
```

With new-env, we have a type-graph like

```
0 ~> t_i_from_func_sig = t_i_param ~> number
```

where `t_i_from_func_sig= t_i_param` is established during `Func_sig.toplevels`. Note that we do have different tvars for parameters (which exist as an entry in the env), and for function signatures. In non-resolved-env setup, the tvar in function signatures is the tvar that first interacts with the optional annotation on function parameters, and got unified with the parameter tvar during `Func_sig.toplevels` with a bind.

In resolved-env when we consider the function to be "fully annotated", `Func_sig.toplevels` won't be called and the link `t_i_from_func_sig = t_i_param` will be broken. For functions that are actually fully annotated or in the future functions parameters that can be contextually typed, it won't be a problem. However, before contextual typing is ready, we need to reestablish the link.

Reviewed By: jbrown215

Differential Revision: D37216078

fbshipit-source-id: 6a14c56df22b28721259576d5e793a27d99bfb6c