[new-env] Unbreak the cycle: create a write entry for class self type
Summary:
Changelog: [internal]

## Motivations

1. Get rid of one of the last few `Flow.unify` for new-env in statement.ml
2. Replace tvar-based self-reference handling with an env-entry based approach, which makes it more compatible with LTI.
3. To add in the `this` type for substitution map for fully resolved environment out of order, we need to have the ability to query the env for the type of a class (it's ok if it's an OpenT).

## Self-reference in existing code

Almost all of self reference in JS definitions needs an named entity so that the code has *something* to self reference. There is one exception: nameless classes. Although there is no syntax to self-reference a nameless class, self-reference still exists at type level due to the modeling of this type. For example,

```
class Foo {
  foo(): this
}
```
is modeled like the following in Flow.

```
class Foo<this: Foo> {
  foo(): this
}
```

Although this doesn't really work for nameless classes. Therefore, for all classes, we create a `self` tvar for self reference for `this` type.

## New approach

This diff changes the modeling for new-env to make it more like

```
type T = class_sig <this: T> {
  foo(): this
}
```

In the implementation, of course, we don't give the class sig binding any meaning name, but instead it's modeled as a special write in name_resolver: it's a non-syntactic write with no syntactic read, but it will be read during type-level operations (for this typing). Now the fully-resolved environment can also query the env-entry to compute this type out of order.

Reviewed By: jbrown215

Differential Revision: D37199436

fbshipit-source-id: 6e80208f24ecb165f3d18cb88aab2ba8ea7d0204